+++ src/Exception/NetworkException.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * OData client library
+ *
+ * @author  Михаил Красильников <m.krasilnikov@yandex.ru>
+ * @license MIT
+ */
+namespace Mekras\OData\Client\Exception;
+
+/**
+ * Network exception.
+ *
+ * @since 0.3
+ */
+class NetworkException extends RuntimeException
+{
+}